+++ security/__init__.py
@@ -2,7 +2,6 @@
 
 from .auth_service import SecurityService
 from core.input_validation import InputValidator, Validator
-from .file_validator import SecureFileValidator
 from .dataframe_validator import DataFrameSecurityValidator
 from .sql_validator import SQLInjectionPrevention, SQLSecurityConfig
 from .xss_validator import XSSPrevention
@@ -16,7 +15,6 @@ __all__ = [
     "SecurityService",
     "InputValidator",
     "Validator",
-    "SecureFileValidator",
     "DataFrameSecurityValidator",
     "SQLSecurityConfig",
     "SQLInjectionPrevention",
